Move 'BlockchainParameters' from Primitive.Model -> Primitive.Types
1. No reasons it's define in the Model rather than the types (although we should really break this gigantic .Types into smaller pieces)
2. Without this, finding a good home for 'ActiveSlotCoefficient' would have been awkward
